pyproject.toml
[tool.poetry]
name = "azure-ai-tooling"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
azure-identity = "^1.15.0"
azure-keyvault-secrets = "^4.8.0"
pydantic = "^2.6.4"
opencensus-ext-azure = "^1.1.13"
azureml-core = "^1.55.0.post2"
azure-mgmt-monitor = "^6.0.2"
azure-monitor-query = "^1.2.1"
azure-cosmos = "^4.6.0"
azure-storage-blob = "^12.19.1"
langchain = "^0.1.12"
azure-search-documents = "^11.4.0"
faker = "^24.3.0"
azure-functions = "^1.18.0"
langchain-community = "^0.0.29"
beautifulsoup4 = "^4.12.3"
moviepy = "^1.0.3"
azure-cognitiveservices-speech = "^1.36.0"
azure-devops = "^7.1.0b4"


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.9"
black = "^22.12.0"
pytest = "^7.4.3"
isort = "^5.13.2"
pytest-asyncio = "^0.23.6"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.3"
poetry-plugin-export = "^1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q -s"
testpaths = ["tools",]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = ["async: marks tests as async (select with '-m async')", "slow: marks tests as slow (select with '-m slow')"]


[tool.pylint]
fail-under = 8.5
max-line-length = 100
max-complexity=10
good-names = ["i", "j", "k", "ex", "Run", "_"]
disable = ["C0114", "C0115", "C0116", "C0301", "W3101"]
load-plugins = ["pylint.extensions.mccabe", ]


[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '(\.git | \.hg | \.mypy_cache | \.tox | \.venv | _build | buck-out | build | dist )/'


[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100
